Crate plasm_runtime

The Substrate runtime. This can be compiled with ``#[no_std]`, ready for Wasm.

Modules

api

constants
Constant values used within the runtime. A set of constant values used in substrate runtime.

impls
Implementations of some helper traits passed into runtime modules as associated types. Some configurable implementations as associated type for the plasm runtime.

Structs

AvailableBlockRatio

BlockHashCount

BondingDuration

ContractCreationFee

ContractFee

ContractTransactionBaseFee

ContractTransactionByteFee

ContractTransferFee

CreationFee

EpochDuration

ExistentialDeposit

ExpectedBlockTime

GenesisConfig

MaximumBlockLength

MaximumBlockWeight

MinimumPeriod

ModuleToIndex
Provides an implementation of ModuleToIndex to map a module to its index in the runtime.

RentByteFee

RentDepositOffset

ReportLatency

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

SessionKeys

SessionsPerEra

SurchargeReward

TargetBlockFullness

TombstoneDeposit

TransactionBaseFee

TransactionByteFee

TransferFee

Version

WeightFeeCoefficient

WindowSize

Enums

BalancesCall
Dispatchable calls.

Call

Event

Forcing
Mode of era-forcing.

Origin

TimestampCall
Dispatchable calls.

Constants

VERSION
Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage
Complex storage builder stuff.

Functions

native_version
Native version.

Type Definitions

Address
The address format for describing accounts.

Babe

BabeConfig

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Contracts

ContractsConfig

Executive
Executive: handles dispatch to the various modules.

FinalityTracker

Gas

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Indices

IndicesConfig

Operator

PlasmStaking

PlasmStakingConfig

RandomnessCollectiveFlip

Session

SessionConfig

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

Trading

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.