Struct plasm_runtime::impls::LinearWeightToFee

pub struct LinearWeightToFee<C>(_);

Convert from weight to balance via a simple coefficient multiplication The associated type C encapsulates a constant in units of balance per weight