Crate plasm_runtime

The Substrate runtime. This can be compiled with ``#[no_std]`, ready for Wasm.

Modules

api

constants
Constant values used within the runtime. A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

BlockGasLimit

BlockHashCount

BondingDuration

ChainId

DeletionQueueDepth

DeletionWeightLimit

DepositPerContract

DepositPerStorageByte

DepositPerStorageItem

DisabledValidatorsThreshold

EcdsaUnsignedPriority

EpochDuration

EthereumFindAuthor

ExistentialDeposit

ExpectedBlockTime

GasWeightMapping

GenesisConfig

ImOnlineUnsignedPriority

IndexDeposit

MaxCodeSize

MaxDepth

MaxLocks

MaxNickLength

MaxScheduledPerBlock

MaxValueSize

MaximumSchedulerWeight

MinNickLength

MinVestedTransfer

MinimumMultiplier

MinimumPeriod

NickReservationFee

OffencesWeightSoftLimit

Origin

PalletInfo
Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RentFraction

ReportLongevity

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionDuration

SessionKeys

SessionsPerEra

SignedClaimHandicap

SurchargeReward

TargetBlockFullness

TombstoneDeposit

TransactionByteFee

TransactionConverter

UncleGenerations

Version

Enums

BalancesCall

Call

Event

OriginCaller

TimestampCall

Constants

GAS_PER_SECOND
Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU). Given the 500ms Weight, from which 75% only are used for transactions, the total EVM execution gas limit is: GAS_PER_SECOND * 0.500 * 0.75 => 6_000_000.

VERSION
Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_GAS
Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Traits

BuildStorage
Complex storage builder stuff.

Functions

native_version
Native version.

Type Definitions

Address
The address format for describing accounts.

AllModules
All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem
All pallets included in the runtime as a nested tuple of types.

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block
Block type as expected by this runtime.

BlockId
BlockId type as expected by this runtime.

CheckedExtrinsic
Extrinsic type that has already been checked.

Contracts

ContractsConfig

EVM

EVMConfig

EthCall

Ethereum

EthereumConfig

Executive
Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header
Block header type as expected by this runtime.

Historical

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Nicks

Offences

Operator

PlasmRewards

PlasmRewardsConfig

PlasmValidator

PlasmValidatorConfig

RandomnessCollectiveFlip

Scheduler

Session

SessionConfig

SignedBlock
A Block signed with a Justification

SignedExtra
The SignedExtension to the basic transaction logic.

SignedPayload
The payload being signed in transactions.

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic
Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig