Crate plasm_runtime

The Substrate runtime. This can be compiled with ``#[no_std]`, ready for Wasm.

Modules

api

constants

Constant values used within the runtime. A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

BasicDeposit

BlockGasLimit

BlockHashCount

BondingDuration

ChainId

DeletionQueueDepth

DeletionWeightLimit

DepositPerContract

DepositPerStorageByte

DepositPerStorageItem

DisabledValidatorsThreshold

EcdsaUnsignedPriority

ElectionLookahead

EpochDuration

EthereumFindAuthor

ExistentialDeposit

ExpectedBlockTime

FieldDeposit

FixedGasPrice

GasWeightMapping

GenesisConfig

ImOnlineUnsignedPriority

IndexDeposit

MaxAdditionalFields

MaxCodeSize

MaxDepth

MaxIterations

MaxLocks

MaxNickLength

MaxNominatorRewardedPerValidator

MaxRegistrars

MaxScheduledPerBlock

MaxSubAccounts

MaxValueSize

MaximumSchedulerWeight

MinNickLength

MinSolutionScoreBump

MinVestedTransfer

MinimumMultiplier

MinimumPeriod

NickReservationFee

OffencesWeightSoftLimit

Origin

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RentFraction

ReportLongevity

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionDuration

SessionKeys

SessionsPerEra

SignedClaimHandicap

SlashDeferDuration

StakingUnsignedPriority

We prioritize im-online heartbeats over election solution submission.

SubAccountDeposit

SurchargeReward

TargetBlockFullness

TombstoneDeposit

TransactionByteFee

TransactionConverter

UncleGenerations

Version

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

Enums

BalancesCall

Call

Event

OriginCaller

StakerStatus

Indicates the initial status of the staker.

TimestampCall

Constants

GAS_PER_SECOND

Current approximation of the gas/s consumption considering EVM execution over compiled WASM (on 4.4Ghz CPU). Given the 500ms Weight, from which 75% only are used for transactions, the total EVM execution gas limit is: GAS_PER_SECOND * 0.500 * 0.75 => 40_000_000.

VERSION

Runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

WEIGHT_PER_GAS

Approximate ratio of the amount of Weight per Gas. u64 works for approximations because Weight is a very small unit compared to gas.

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

Native version.

Type Definitions

Address

The address format for describing accounts.

AllModules

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem

All pallets included in the runtime as a nested tuple of types.

Authorship

Babe

BabeConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

Contracts

ContractsConfig

EVM

EVMConfig

EthCall

Ethereum

EthereumConfig

Executive

Executive: handles dispatch to the various modules.

Grandpa

GrandpaConfig

Header

Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

Indices

IndicesConfig

Nicks

Offences

RandomnessCollectiveFlip

Scheduler

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

SignedPayload

The payload being signed in transactions.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

Vesting

VestingConfig